requirements/base.txt
scikit-learn
pandas
joblib
pydantic
pyyaml